Wrote interface and simple class for query processing + test
As queries can be processed in many ways (e.g. by using other tools
to form the question "where is the birthplace of bill gates wife" to
the running example "birthplace bill gates wife"), an interface is given.

SimpleQueryProcessing is the easiest form, taking the query and
directly giving it to the NGramHierarchy-class.