Swap position of length and index in all NGramHierarchy-methods
This change streamlines the implementation, because it is already used in
NGramEntryPosition. Furthermore it feels more natural to first give the
length of the n-gram (and therefore providing the information in which depth
of the n-gram hierarchy we are) and the index of the n-gram within all
n-grams of the same depth.