Move query process definition to constructor
The query processing should be the same in every instance, so
a new query processing definition in each answer-call is not efficient.
While it should not make a real difference for now, later it can be expected
to have more advanced query processing classes which work with dictionaries or web services.