Add support for separate file handling
(The changes in TsvDictionaryImport were made for testing purposes of TsvFileHandler)

Previously we had to implement a new class that handles the file AND builds the dictionary for each type of file. Now the file handling is separated.
This will be handy in the future, where we probably want to use other technologies than in-memory-HashMaps to hold the dictionary.