Remove "must be new"-restriction
First commit to fix #11.
Now, already existing nodes can be handled as "missing triplet".
This means that 3 already existing nodes in a graph can interlink with each other (by adding a fact node).

This commit creates a new bug, because the graphs cannot recognize two nodes with the same content as the same node at the moment.