Rework updating explanation score
Tests after the last commit indicated that the explanation score was calculated wrongly.

This commit reworked updating the explanation score.
It now takes the colors of a node. If new colors are added, the length of the NGramEntryPosition are added to the explanation score. Because the length indicates how long the n-gram is, the reworked explanation score should be correct as long as every node has the correct colors.