Rework rdf:type-answer processing
Previously we assumed that the answer is always an instance of a class.
However, it can be expected that questions like "type of Lion King" can occure (especially if a pre-processor parses questions like "What is Lion King=").
In this case we want the class instead of the instace.
This implementation is a rework considers this fact.